requirements.txt
actions==0.1.0
adbutils==1.2.9
aiofiles==23.1.0
annotated-types==0.5.0
anyio==3.6.2
apkutils2==1.0.0
asgiref==3.6.0
av==10.0.0
cached-property==1.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
cigam==0.0.3
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
decorator==5.1.1
deprecated==1.2.13
deprecation==2.1.0
filelock==3.12.0
flatbuffers==23.5.9
h11==0.14.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
inflection==0.5.1
jellyfish==0.11.2
logzero==1.7.0
lxml==4.9.2
lz4==4.3.2
markdown-it-py==2.2.0
mdurl==0.1.2
mpmath==1.3.0
numpy==1.24.3
onepush==1.3.0
onnxruntime==1.14.1
opencv-python==4.7.0.72
packaging==20.9
pillow==9.5.0
pponnxcr==2.0
prettytable==2.2.1
progress==1.6
protobuf==4.23.0
psutil==5.9.3
py==1.11.0
pyclipper==1.3.0.post4
pycryptodome==3.20.0
pydantic==2.4.2
pydantic-core==2.10.1
pyelftools==0.29
pygments==2.15.1
pyparsing==3.0.9
pyreadline3==3.4.1
python-dotenv==1.0.0
pywebio==1.8.3
pyyaml==6.0
requests==2.30.0
retry==0.9.2
rich==13.3.5
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.14.2
sympy==1.12
tornado==6.3.1
tqdm==4.65.0
typing-extensions==4.8.0
ua-parser==0.16.1
uiautomator2==2.16.17
uiautomator2cache==0.3.1
urllib3==2.0.2
user-agents==2.2.0
uvicorn[standard]==0.17.6
watchgod==0.8.2
wcwidth==0.2.6
websockets==11.0.3
whichcraft==0.6.1
wrapt==1.15.0
xmltodict==0.13.0